+++ config.inc.dev.php
@@ -2,11 +2,6 @@
 
 /* If you make changes here check if changes at the production server are needed */
 
-if (php_sapi_name() != 'cli') {
-    /* Whoops catches all error messages in CLI mode as well :( */
-    Foodsharing\Debug\Whoops::register();
-}
-
 $protocol = 'http';
 if (isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] == 'on') {
     $protocol = 'https';