+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - removed CHANGELOG.md merge=union !2866 @chriswalg
 - Updated quality tools !2867 @\_fridtjof_
 - Show CS Fixer results in gitlab !2867 @\_fridtjof_
+- Added the Symfony profiler and debug toolbar !1580 @\_fridtjof
 
 ## Development documentation